update: dashboard route home
