+++ frontend/src/App.tsx
@@ -1,5 +1,6 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { LoginPage } from "./pages/LoginPage.tsx";
+import { Home } from "./pages/Home.tsx";
 import { DashboardPage } from "./pages/DashboardPage.tsx";
 import { RegisterPage } from "./pages/RegisterPage.tsx";
 
@@ -9,8 +10,10 @@ function App() {
       <Routes>
         <Route path="/" element={<LoginPage />} />
         <Route path="/login" element={<LoginPage />} />
-        <Route path="/dashboard" element={<DashboardPage />} />
         <Route path="/register" element={<RegisterPage />} />
+        <Route path="/dashboard" element={<DashboardPage />}>
+          <Route path="home" element={<Home />} />
+        </Route>
       </Routes>
     </Router>
   );
