+++ frontend/src/pages/DashboardPage.tsx
@@ -4,9 +4,7 @@ import { ThemeToggleButton } from "../components/ThemeToggleButton";
 // import RoomsSection from './RoomsSection';
 // import UsersSection from './UsersSection';
 // import EventsSection from './EventsSection';
-import { Statistics } from "../components/Statistics";
-import { RecentActivities } from "../components/RecentActivities";
-import { useNavigate } from "react-router-dom";
+import { Outlet, useNavigate } from "react-router-dom";
 import { Navbar } from "../components/Navbar";
 
 export const DashboardPage = () => {
@@ -21,29 +19,6 @@ export const DashboardPage = () => {
   // const handleSectionChange = (newSection) => {
   //   setSection(newSection);
   // };
-
-  // Fictitious data for RecentActivities component
-  const activities = [
-    {
-      user: "John Doe",
-      room: "Conference Room",
-      event: "Scheduled Meeting",
-      date: new Date(2023, 4, 20, 10, 30),
-    },
-    {
-      user: "Jane Smith",
-      room: "Break Room",
-      event: "Coffee Break",
-      date: new Date(2023, 4, 20, 11, 0),
-    },
-    {
-      user: "Bob Johnson",
-      room: "Office #2",
-      event: "Daily Work",
-      date: new Date(2023, 4, 20, 9, 0),
-    },
-    // ... more activities ...
-  ];
 
   interface ThemeColors {
     [key: string]: string;
@@ -75,16 +50,8 @@ export const DashboardPage = () => {
 
       <div className="dashboard-body flex">
         <Navbar />
-        <div className="dashboard-content flex-grow p-4">
-          <div className="statistics flex flex-col space-y-4 lg:space-y-0 lg:flex-row lg:space-x-4 my-4">
-            <Statistics title="Current Rooms in Use" value={42} />
-            <Statistics title="Events Occurred" value={122} />
-            <Statistics title="Number of Users" value={61} />
-          </div>
-          <div className="my-4">
-            <RecentActivities activities={activities} />
-          </div>
-        </div>
+
+        <Outlet />
       </div>
     </div>
   );
